requirements.txt
discord.py==2.0.0
dnspython
git+https://github.com/Gorialis/jishaku
asyncpg
asyncio
Pillow